+++ src/frontend/Layout.h
@@ -24,7 +24,6 @@
 #include <QString>
 #include <QVector>
 #include <QJsonObject>
-#include "FileSystem.h"
 
 /* Layout handiling code for the frontend */
 
@@ -59,8 +58,6 @@ class Layout {
   LayoutDesc lD;
   /* Internal map of available layouts */
   QMap<QString, QString> layoutMap;
-
-  UserFolders folders;
 
   /* Load Layout Description. Used internaly */
   void loadDesc();